+++ packages/core/src/actions/casino/cointoss.ts
@@ -13,7 +13,7 @@ import type { BetSwirlWallet } from "../../provider";
 import {
   type CasinoBetParams,
   type CasinoPlaceBetOptions,
-  type CasinoPlacedBet,
+  type NormalCasinoPlacedBet,
   type PlaceBetCallbacks,
   getPlacedBetFromReceipt,
   placeBet,
@@ -23,9 +23,10 @@ export interface CoinTossParams extends CasinoBetParams {
   face: COINTOSS_FACE;
 }
 
-export interface CoinTossPlacedBet extends CasinoPlacedBet {
+export interface CoinTossPlacedBet extends NormalCasinoPlacedBet {
   face: COINTOSS_FACE;
   encodedFace: CoinTossEncodedInput;
+  game: CASINO_GAME_TYPE.COINTOSS;
 }
 
 export async function placeCoinTossBet(
@@ -102,6 +103,7 @@ export async function getCoinTossPlacedBetFromReceipt(
   const { args } = decodedCoinTossPlaceBetEvent;
   return {
     ...gamePlacedBet,
+    game: CASINO_GAME_TYPE.COINTOSS,
     encodedFace: args.face,
     face: CoinToss.decodeInput(args.face),
   };