+++ packages/core/src/actions/index.ts
@@ -4,6 +4,8 @@ export * from "./casino/dice";
 export * from "./casino/roulette";
 export * from "./casino/keno";
 export * from "./casino/game";
+export * from "./casino/wheel";
+export * from "./casino/weightedGame";
 
 /* Common */
 export * from "./common/approve";