+++ packages/core/src/utils/bet.ts
@@ -1,9 +1,11 @@
 import { BP_VALUE } from "../constants";
-import { CASINO_GAME_TYPE, type ChainId, slugById } from "../data";
+import { CASINO_GAME_TYPE, type ChainId, type NORMAL_CASINO_GAME_TYPE, slugById } from "../data";
 import { CoinToss } from "../entities/casino/cointoss";
 import { Dice } from "../entities/casino/dice";
 import { Keno } from "../entities/casino/keno";
 import { Roulette } from "../entities/casino/roulette";
+import { WeightedGame } from "../entities/casino/weightedGame";
+import type { WeightedGameConfiguration } from "../read";
 
 //houseEdge is in BP_VALUE
 export function getBetSwirlFees(payout: bigint, houseEdge: number): bigint {
@@ -33,7 +35,7 @@ export function getFormattedNetMultiplier(multiplier: number, houseEdge: number)
   return Number((getNetMultiplier(multiplier, houseEdge) / BP_VALUE).toFixed(3));
 }
 
-export function decodeCasinoInput(encodedInput: string, game: CASINO_GAME_TYPE): any {
+export function decodeNormalCasinoInput(encodedInput: string, game: NORMAL_CASINO_GAME_TYPE): any {
   switch (game) {
     case CASINO_GAME_TYPE.DICE:
       return Dice.decodeInput(encodedInput);
@@ -45,7 +47,19 @@ export function decodeCasinoInput(encodedInput: string, game: CASINO_GAME_TYPE):
       return Keno.decodeInput(encodedInput);
   }
 }
-export function decodeCasinoRolled(encodedRolled: string, game: CASINO_GAME_TYPE): any {
+
+// houseEdge is in BP_VALUE
+export function decodeWeightedCasinoInput(
+  encodedInput: string,
+  _weightedGameConfiguration: WeightedGameConfiguration,
+  _houseEdge = 0,
+) {
+  return WeightedGame.decodeInput(encodedInput);
+}
+export function decodeNormalCasinoRolled(
+  encodedRolled: string,
+  game: NORMAL_CASINO_GAME_TYPE,
+): any {
   switch (game) {
     case CASINO_GAME_TYPE.DICE:
       return Dice.decodeRolled(encodedRolled);
@@ -56,6 +70,14 @@ export function decodeCasinoRolled(encodedRolled: string, game: CASINO_GAME_TYPE
     case CASINO_GAME_TYPE.KENO:
       return Keno.decodeRolled(encodedRolled);
   }
+}
+// houseEdge is in BP_VALUE
+export function decodeWeightedCasinoRolled(
+  encodedRolled: string,
+  weightedGameConfiguration: WeightedGameConfiguration,
+  houseEdge = 0,
+): any {
+  return WeightedGame.decodeRolled(encodedRolled, weightedGameConfiguration, houseEdge);
 }
 
 export function formatChainlinkSubscriptionUrl(subscriptionId: string | bigint, chainId: ChainId) {