+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @betswirl/sdk-core
 
+## 0.1.0
+
+### Minor Changes
+
+- Add Wheel game with weighted gamr logic
+
 ## 0.0.9
 
 ### Patch Changes